Adjustments to `board.go` file
Param `position` in the `insert` method was renamed to `pos` to be more consistent with rest of the codebase.

Added `check` method to abstract both checks on coordinates before inserting.

`checkCoordsDomain` moved from `coords.go` file and turned into function which takes in `pos` parameter.

Added `checkEmpty` method which checks if the position is empty.